requirements.txt
supabase
python-dotenv
requests
feedparser
flask
google-generativeai
